refacto: clean up and polish the proximity analysis service
With a teeny-tiny improvement on performance, because it was just hanging there.
